services/model-generator/requirements.txt
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-instrumentation-fastapi==0.60b1
fastapi==0.128.0
trimesh==4.11.0
pydantic==2.12.5
pydantic-settings==2.12.0
asyncpg==0.31.0
boto3==1.42.24
structlog==25.5.0
fastmcp==2.14.2
mcp==1.25.0
aiofiles==25.1.0
taskiq-fastapi==0.4.0
taskiq-redis==1.2.1
taskiq==0.12.1